feat/#21: 카테고리 목록 불러오기 및 카테고리 추가 API 구현
