+++ app/src/main/AndroidManifest.xml
@@ -40,7 +40,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".view.home.MainActivity"
+        <activity android:name=".view.MainActivity"
             android:exported="false">
         </activity>
