+++ app/src/main/java/com/yeolsimee/moneysaving/di/DataModule.kt
@@ -1,11 +1,14 @@
 package com.yeolsimee.moneysaving.di
 
+import com.yeolsimee.moneysaving.data.api.CategoryApiService
 import com.yeolsimee.moneysaving.data.api.RoutineApiService
 import com.yeolsimee.moneysaving.data.api.UserApiService
+import com.yeolsimee.moneysaving.data.repository.CategoryApiRepository
 import com.yeolsimee.moneysaving.data.repository.RoutineApiRepository
 import com.yeolsimee.moneysaving.data.repository.UserApiRepository
 import com.yeolsimee.moneysaving.data.source.RoutineSource
 import com.yeolsimee.moneysaving.data.source.UserSource
+import com.yeolsimee.moneysaving.domain.repository.ICategoryApiRepository
 import com.yeolsimee.moneysaving.domain.repository.IRoutineApiRepository
 import com.yeolsimee.moneysaving.domain.repository.IUserApiRepository
 import dagger.Module
@@ -33,4 +36,8 @@ object DataModule {
     @Provides
     @Singleton
     fun provideRoutineSource(api: RoutineApiService): RoutineSource = RoutineSource(api)
+
+    @Provides
+    @Singleton
+    fun provideCategoryApiRepository(api: CategoryApiService): ICategoryApiRepository = CategoryApiRepository(api)
 }
