+++ data/src/main/java/com/yeolsimee/moneysaving/data/NetModule.kt
@@ -3,6 +3,7 @@ package com.yeolsimee.moneysaving.data
 import android.app.Application
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
+import com.yeolsimee.moneysaving.data.api.CategoryApiService
 import com.yeolsimee.moneysaving.data.api.RoutineApiService
 import com.yeolsimee.moneysaving.data.api.UserApiService
 import com.yeolsimee.moneysaving.data.interceptor.ConnectivityInterceptor
@@ -77,6 +78,12 @@ class NetModule {
     @Singleton
     fun provideRoutineApiService(retrofit: Retrofit): RoutineApiService {
         return retrofit.create(RoutineApiService::class.java)
+    }
+
+    @Provides
+    @Singleton
+    fun provideCategoryApiService(retrofit: Retrofit): CategoryApiService {
+        return retrofit.create(CategoryApiService::class.java)
     }
 
     companion object {
