+++ domain/src/main/java/com/yeolsimee/moneysaving/domain/entity/category/CategoryNameRequest.kt
@@ -1,0 +1,5 @@
+package com.yeolsimee.moneysaving.domain.entity.category
+
+data class CategoryNameRequest(
+    val categoryName: String
+)
