+++ domain/src/main/java/com/yeolsimee/moneysaving/domain/entity/routine/Routine.kt
@@ -5,8 +5,8 @@ data class Routine(
     val routineName: String,
     val routineCheckYN: String,
     val routineTimeZone: String,
-    val alarmTimeHour: String,
-    val alarmTimeMinute: String
+    val alarmTimeHour: String = "",
+    val alarmTimeMinute: String = ""
 ) {
     fun getTimeZoneText(): String {
         return when (routineTimeZone) {
