+++ domain/src/main/java/com/yeolsimee/moneysaving/domain/repository/ICategoryApiRepository.kt
@@ -1,0 +1,12 @@
+package com.yeolsimee.moneysaving.domain.repository
+
+import com.yeolsimee.moneysaving.domain.entity.category.TextItem
+
+interface ICategoryApiRepository {
+    suspend fun getCategoryList(
+    ): Result<MutableList<TextItem>>
+
+    suspend fun addCategory(
+        name: String
+    ): Result<TextItem>
+}
